refactor: ui 패키지 이름 controller 로 수정
